llvm: Update to 15.0.7, cleanup the packaging (#9)
Our future goal is to split out the LLVM packaging across smaller
more maintainable units, and this is our first step towards that.

As we're now building with the system clang from the last package
build, the bootstrap step has been removed. If a rebootstrap is
needed in future, we can do so via `bill` and rebuild the distro
properly.

Additionally the cleaned packaging is now validated to provide
clang-32bit to enable a multilib LLVM toolchain.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>