llvm: enable AMDGPU and keep static libraries
Both changes shouldn't break anything.
- Enable AMDGPU backend.
- Keep `libclang*.a` static libraries in the `clang-devel` package.

While I agree it's generally desired to remove static libraries, in this
case it is not because it will break CMake detection. For example, when
a certain package depends on the `Clang::Clang` target, CMake expects
those `libclang*.a` static libraries to exist even though the actual
package may never link those libraries.

Unless we can find a way to patch LLVM or CMake to not bother with
these missing static libraries, I suggest that we keep the static
libraries for now. Since `clang-devel` rarely gets pulled in, this
shouldn't add a build significant time for most packages.

Signed-off-by: Gavin Zhao <git@gzgz.dev>