libs/fonts/graphite: Fix up linkage and enable emul32 build
We set the dependency to a binary one for readelf, and remove the ugly hacks that
forced linking to `libgcc_s` so that we can retain a pure LLVM build. On top of
that, we use the legacy `LIB_DIR` macro (undocumented in tree) to ensure we can
build an emul32 build.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>