systemd: Statically define base groups for legacy tooling
Our intent in Serpent OS is to eventually deprecate the use of `/etc/passwd`
and similar files for system group and user definitions. However, the userspace
tooling is very tied into these locations.

For example, shadow, accountsservice, and many shell scripts expect to
be able to interact directly with these files, rather than using the proper
APIs that interact with NSS.

We also note this is holding back desktop environmens like GNOME, and have
it on our radar to assist in unblocking this situation. One consideration
is synthetic etc files symlinked from `/run` that are dynamically handled
via FUSE.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>