systemd: Add binary(setfont) as a rundep
At boot, systemd starts the vconsole service to set up legible default
console fonts in both the initrd and in the real system root.

For this to work, the /usr/bin/setfont binary needs to be available,
which is part of `kbd`.

Therefore, add binary(setfont) as a systemd rundep so we can get nice
looking fonts during and after boot in the linux console.

Signed-off-by: Rune Morling <ermo@serpentos.com>