ncurses: Fix pkgconfig
Ncurses was leaking the LDFlags into the pkgconfig file. Boulder adds `--icf=safe` by default when building with LLVM, which was getting put into the flags in the pkgconfig file. Since binutils doesn't support this flag that caused any gnu toolchain builds that tried to link against ncurses to fail.

Also, switch the C++ bindings to be a shared library

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>