cryptsetup: Update to v2.7.4
Packaging updates:
Ccopy src/*setup to /usr/bin/ as the install script doesn't
appear to do that by default, despite the fact that the binaries appear
to be built by default.

Should probably create a patch and upstream it but I cba to do that rn.

Release Notes:
- [2.7.1](https://mirrors.edge.kernel.org/pub/linux/utils/cryptsetup/v2.7/v2.7.1-ReleaseNotes)
- [2.7.2](https://mirrors.edge.kernel.org/pub/linux/utils/cryptsetup/v2.7/v2.7.2-ReleaseNotes)
- [2.7.3](https://mirrors.edge.kernel.org/pub/linux/utils/cryptsetup/v2.7/v2.7.3-ReleaseNotes)
- [2.7.4](https://mirrors.edge.kernel.org/pub/linux/utils/cryptsetup/v2.7/v2.7.4-ReleaseNotes)

Git Changelog:
- https://gitlab.com/cryptsetup/cryptsetup/-/compare/v2.7.0...v2.7.4

Signed-off-by: Rune Morling <ermo@serpentos.com>