justfile: Add {create,index,ls,mv}-local targets
All targets use the LOCAL_REPO environment variable.

The default LOCAL_REPO location is set to:

    ${HOME}/.cache/local_repo/x86_64

for XDG compliance.

This can trivially be overridden in the recipes/.env file or via an
environment variable, which is expected to be useful for stack work.

If the LOCAL_REPO dir does not exist, it will be created first. This is
also intended to be useful for stack work.

Note that the options for the ls-local target file listing have been
carefully chosen so the list is sorted by file creation time, which is
useful for checking if things have been built in the correct order and
whether all files present have been indexed.

Signed-off-by: Rune Morling <ermo@serpentos.com>