gnome-session: Fix xwayland integration again
For some unbeknown reason, xwayland is lumbered into general "x11"
support in GNOME Session. This means that even if we disable the X11
session, Xwayland supporting services are not spawned.

To rectify this, we enable X11 support and nuke the X11-specific login
session to keep a Wayland session with Xwayland support. Phew!

Signed-off-by: Ikey Doherty <ikey@serpentos.com>