llvm: Fix various compiler crashes for the latest kernel
The most important fix for us here being fixing a crash in the compiler
when encountering `strucft net_device` in `net/core/dev.c`, which has
the 6.11 kernel series failing to build from source.

This is fixed in LLVM19, but we're backporting these fixes into our LLVM18
due to some further gotchas with amdgpu and some programming languages in
the newer series.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>