coreutils: Update to v0.0.28+git.a644724
This is to address a permissions issue with `install` seen in
kernel header packaging.

We also go to some decent lengths now to enforce use of our own cargo build
macros, bypassing the makefile, and ensuring proper distribution integration.
Atop that, we now build all of the manpages and completions that were missing.

Signed-off-by: Ikey Doherty <ikey@serpentos.com>