golang: Fix goroot auto-detection
Having GOROOT set during the build causes said value to be emitted into the binary symbol table, which can apparently cause the automatic detection of GOROOT by resolving the /usr/bin/go symlink to fail. Ensure that it is unset during the build and now that it's no longer necessary we can remove the profile.d script setting it.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>